refactor(examples): cleanups & tweaks
